Parse multiple 'checks=...' entries
Previously, 'Checks' kept 'string', which was a copy of the one that's
given from the compiler by 'ParseArgs' call. Then an internal
'unordered_map<string_view, ...>' referred to it. Since I couldn't find
any modification of 'compilerInstance.getFrontendOpts().PluginArgs' in
the Clang sources, I believe, the map can refer to the strings given by
the compiler itself and remove copy of the plugin arguments. This is the
commit to suspect if arguments parsing blows up.

Since the map and the
string were cleaned at each 'Checks::parse' call, we could only keep
track of the last 'checks=...' argument. It was decided that we should
consider all the arguments of this form.